+++ fabric/src/main/java/dev/nyon/magnetic/mixins/blocks/ShulkerBoxBlockMixin.java
@@ -1,0 +1,37 @@
+package dev.nyon.magnetic.mixins.blocks;
+
+import com.llamalad7.mixinextras.injector.v2.WrapWithCondition;
+import dev.nyon.magnetic.utils.MixinHelper;
+import net.minecraft.core.BlockPos;
+import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.item.ItemEntity;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.ShulkerBoxBlock;
+import net.minecraft.world.level.block.state.BlockState;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+
+@Mixin(ShulkerBoxBlock.class)
+public class ShulkerBoxBlockMixin {
+
+    @WrapWithCondition(
+        method = "playerWillDestroy",
+        at = @At(
+            value = "INVOKE",
+            target = "Lnet/minecraft/world/level/Level;addFreshEntity(Lnet/minecraft/world/entity/Entity;)Z"
+        )
+    )
+    private boolean useMagneticInsteadOfDrop(
+        Level instance,
+        Entity entity,
+        Level world,
+        BlockPos pos,
+        BlockState state,
+        Player player
+    ) {
+        if (!(player instanceof ServerPlayer serverPlayer)) return true;
+        return MixinHelper.wrapWithConditionPlayerItemSingle(serverPlayer, ((ItemEntity) entity).getItem());
+    }
+}